Rename resource edit forms
