+++ src/components/resource/edit-custom-title/index.js
@@ -1,0 +1,1 @@
+export { default } from './resource-edit-custom-title';
