--- src/components/resource/edit-custom/index.js
@@ -1,1 +1,0 @@
-export { default } from './custom-resource-edit';
