+++ src/components/resource/edit-managed-title/index.js
@@ -1,0 +1,1 @@
+export { default } from './resource-edit-managed-title';
