+++ src/components/resource/edit-managed-title/resource-edit-managed-title.css
