--- src/components/resource/edit-managed/index.js
@@ -1,1 +1,0 @@
-export { default } from './managed-resource-edit';
